added helperFunctions.js and cleaned up dashboard
