+++ src/components/dashboardComponents/ChooseDateForm.js
@@ -7,7 +7,9 @@ import {
   IconButton
 } from '@chakra-ui/core';
 import { useAuth } from '../../contexts/auth';
+import {convertTime} from '../../utils/helperFunctions'
 import axios from 'axios'
+
 
 
 //TODO LINK UP THE DAYS.JS TRIGGER TO MATCH THE TEMPLATE TOGGLE FROM A STATE WIDE SELECTION.
@@ -147,35 +149,6 @@ const ChooseDateForm = ({
   );
 };
 
-
-
-
-    const convertTime = (time)=>{
-      // code converts response.data.starttime to number
-
-      
-      if (time){
-
-          let splitStartTime = time.split(':');
-          let joinStartTime = splitStartTime.join('');
-          let startTimeAsNumber = parseInt(joinStartTime, 10);
-      
-          // fn for converting response.data.starttime and/or endtime back to time string (from number)
-          function convertToTime(value, index) {
-            return value.substring(0, index) + ":" + value.substring(index);
-          }
-      
-          // converts times from 24 hour to 12 hour format
-          if (startTimeAsNumber >= 1300) {
-            startTimeAsNumber -= 1200;
-            let startTimeAsString = startTimeAsNumber.toString();
-            let convertedStartTime = convertToTime(startTimeAsString, startTimeAsString.length - 2);
-            return convertedStartTime + 'pm';
-          } else {
-            return time + 'am';
-          }
-      }
-    }
 
 
 export default ChooseDateForm;
