+++ src/components/mobile/Mobile.js
@@ -8,7 +8,7 @@ import Groups from './Groups'
 
 const Mobile = () => {
     // 0 = calendar, 1 = events, 2 = groups
-    const [navState, setNavState] = useState(2) 
+    const [navState, setNavState] = useState(0) 
     const [formOpen, setFormOpen] = useState(false);
     const [templateFormOpen, setTemplateFormOpen] = useState(false);
     const [conStart, setConStart] = useState("");
