+++ src/components/mobile/eventComponents/MobileChooseDateForm.js
@@ -1,12 +1,8 @@
 import React, {useEffect} from 'react';
-import {
-  Flex,
-  Heading,
-  IconButton
-} from '@chakra-ui/core';
 import { useAuth } from '../../../contexts/auth';
 import axios from 'axios'
 import styled from 'styled-components'
+import {convertTime} from '../../../utils/helperFunctions'
 
 
 const MobileChooseDateForm = ({
@@ -14,27 +10,21 @@ const MobileChooseDateForm = ({
   starttime,
   endtime,
   summary,
-  description,
-  selected,
   setSelected,
-  templateFormOpen,
   setTemplateFormOpen,
   setTemplateList,
   templateList,
   setNavState,
-  formOpen,
   setFormOpen,
   setToggleNav, 
-  toggleNav,
   conStart, 
   setConStart, 
   conEnd, 
   setConEnd, 
-  summ, 
   setSumm
 }) => {
 
-  const { googleApi, api } = useAuth();
+  // const { googleApi, api } = useAuth();
   const deleteTemplate = async id => {
     try {
       const response = await axios.delete(
@@ -60,7 +50,7 @@ const MobileChooseDateForm = ({
   }
 
 
-
+  //sets the summary and time displayed on event page
   useEffect(()=>{
     setSumm(summary);
     if (starttime){
@@ -72,7 +62,7 @@ const MobileChooseDateForm = ({
   },[starttime, endtime])
   
 
-
+  //sets the summary and starttime of the event actually being applied
   const handleCalendarView = () =>{
     setSumm(summary);
     if (starttime){
@@ -106,35 +96,6 @@ const MobileChooseDateForm = ({
   );
 };
 
-
-
-
-    const convertTime = (time)=>{
-      // code converts response.data.starttime to number
-
-      
-      if (time){
-
-          let splitStartTime = time.split(':');
-          let joinStartTime = splitStartTime.join('');
-          let startTimeAsNumber = parseInt(joinStartTime, 10);
-      
-          // fn for converting response.data.starttime and/or endtime back to time string (from number)
-          function convertToTime(value, index) {
-            return value.substring(0, index) + ":" + value.substring(index);
-          }
-      
-          // converts times from 24 hour to 12 hour format
-          if (startTimeAsNumber >= 1300) {
-            startTimeAsNumber -= 1200;
-            let startTimeAsString = startTimeAsNumber.toString();
-            let convertedStartTime = convertToTime(startTimeAsString, startTimeAsString.length - 2);
-            return convertedStartTime + 'pm';
-          } else {
-            return time + 'am';
-          }
-      }
-    }
 
 
 export default MobileChooseDateForm;
